- [Feat] Thêm mới Enum AuctionStatus, tối ưu AuctionSession với ScheduledExecutorService đếm giờ chuẩn xác.
- [Feat] Triển khai PaymentService (Singleton) xử lý thanh toán, giải quyết bài toán chống Race Condition.
- [Merge] Giải quyết dứt điểm xung đột mã nguồn BidTransaction từ nhánh dev, chuẩn hóa an toàn mạng (Serialization).
- [Refactor] Đổi toàn bộ kiểu dữ liệu tiền tệ và số dư từ double sang BigDecimal trong Models.
- [Refactor] Cập nhật logic toán học: Thay thế toán tử bằng compareTo(), add() và subtract() để triệt tiêu sai số làm tròn.

+++ Auction_shared/src/main/java/com/auction/shared/factory/ItemFactory.java
@@ -1,8 +1,10 @@
 package com.auction.shared.factory;
 
+import java.math.BigDecimal;
+
 import com.auction.shared.models.Item;
 
 public interface ItemFactory {
     // tránh tạo 1 item mà thông tin item không minh bạch
-    Item createItem(String name, String description, double startingPrice, double currentPrice);
+    Item createItem(String name, String description, BigDecimal startingPrice, BigDecimal currentPrice);
 }
